requirements.txt
pandas==1.3.5
scikit-learn==1.0.2
apache-airflow-providers-apache-spark